==> Rmd/02-isformswitch.Rmd <==
---
title: "IsoformSwitchAnalzyeR for detecting differential intron retention"
author: "Stuart Lee, Charity Law"
date: "`r Sys.Date()`"
output: 
  bookdown::html_document2:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
                      fig.align = "center",
                      comment = "#>", 
                      message = FALSE,
                      warning = FALSE,
                      fig.path = here::here("img/"),
                      cache.path = here::here("cache/")
                      )


library(plyranges)
library(superintronic)
library(ggplot2)
source(here::here("R", "prettycov.R"))

# external inputs
design <- readRDS(here::here("data", "design.rds"))
parts <- readRDS(here::here("data", "complete-annotation.rds"))
cvg <- readRDS(here::here("data", "complete-coverage.rds"))
all_features <- readRDS(here::here("data", "superintronic-features.rds"))
hits <- readRDS(here::here("data", "superintronic-hits.rds")) 

# update for relevant samples
design <- subset(design, Kit == "polyA")
cvg <- cvg[design$Sample]
```

##  IsoformSwitchAnalyzeR analysis

The Bioconductor package `IsoformSwitchAnalyzeR` also allows for the exploration
of intron retention events using a similar workflow:

1. import isoform quantifications from kallisto/salmon/other program of your choice.

We used kallisto quant on single read mode with an average fragment
size of 175 and a standard deviation of 20, on all fastq files for each
replicate in the polyA kits for each of our cellines. 
We need to use both here since 
IsoformSwitchAnalzyeR requires biological replicates for testing for
switching events.

```{r isf-prepare}
library(IsoformSwitchAnalyzeR)

quants <- importIsoformExpression(
  parentDir = here::here("data", "kallisto/"),
)

```


2. Preparing data structure called a `switchList` by importing annotations
and design matrix.

```{r isf-import}
# use a simplified version of the design
design_isf <- data.frame(sampleID = colnames(quants$abundance[,-1])) %>% 
  mutate(condition =  design$CellLine)

isf <- importRdata(
  isoformCountMatrix = quants$counts,
  isoformRepExpression = quants$abundance,
  designMatrix = design_isf,
  isoformExonAnnoation = here::here("data-raw", "gencode.v27.annotation.gtf.gz"),
  isoformNtFasta = here::here("data-raw", "gencode.v27.transcripts.fa.gz"),
  removeNonConvensionalChr = TRUE,
)

isf 
```


3. Filtering isoforms


Next we prefilter, the isoforms using roughly the same filters we used
in the superintronic analysis above. To speed things up we could further
specify a gene expression filter, but we have left that out for now.

```{r isf-filtered, cahce = TRUE, dependson=isf}
isf_filtered <- preFilter(isf,
                          acceptedGeneBiotype = "protein_coding",
                          removeSingleIsoformGenes = TRUE,
                          geneExpressionCutoff = 1,
                          isoformExpressionCutoff = 0.1,
                          keepIsoformInAllConditions = TRUE
                          )
isf_filtered
```

4. Searching for intron retention events

First we are required to run DEXSeq to test for isoform switching (
this does all pairwise comparisons between conditions, which slows things
down quite a bit.) and between the two kits. We have used quite a small
alpha since DEXSeq seems quite liberal...

```{r isf-dexseq}
isf_analysed <- isoformSwitchTestDEXSeq(isf_filtered, 
                                        reduceToSwitchingGenes = TRUE,
                                        alpha = 0.05,
                                        dIFcutoff = 0.1)
isf_analysed
```


Now we can analyse intron retention events from one kit relative to another:

```{r isf-intron-retention, cache=TRUE}
# this takes quite a while to compute...
isf_analysed <- analyzeIntronRetention(isf_analysed)
isf_analysed <- analyzeSwitchConsequences(isf_analysed, "intron_retention")

# the result
isf_analysed

# isoforms with intron retention events
isf_analysed$switchConsequence %>% 
  dplyr::count(switchConsequence)
```

And the results can be summarised at the gene level and compared to the 
superintronic analysis. 

At the gene level there are around 502 genes declared significant
at 0.01 fdr cut-off between the celllines kits that
are filtered for intron retention events.

```{r top-table}
is_res <- extractTopSwitches(isf_analysed, 
                             filterForConsequences = TRUE, 
                             n = NA) %>% 
  dplyr::as_tibble()

is_res
```

```{r save-res, include = FALSE}
saveRDS(is_res, here::here("data", "isa_hits.rds"))
```

We can also look at the coverage plots for these genes (we will choose
the top 10 genes):

```{r isf-coverage-plots}
top10 <- is_res %>% 
  filter(Rank < 10)

for (i in seq_len(nrow(top10))) {
  input <- top10[i,]
  cplot <- pretty_cov_plot(cvg, parts, input, heights = c(2, 0.25))
  fname <- here::here("img", "isa-cov", 
                      paste0(input$gene_id, "-rank-", sprintf("%02s", input$Rank), ".pdf"))
  ggsave(fname, cplot)
}
```
